merTools: Provides methods for extracting and exploring results from merMod objects in the lme4 package.

The merTools package contains convenience tools for extracting useful information from and exploring the implications of merMod objects created by the lme4 package. These convenience functions are especially useful for merMod objects that take a long time to estimate due to their complexity or because they are estimated on very large samples.

Details

See the vignettes for usage examples

merMod extraction/utility functions

fastdisp

superFactor

REextract

REsim

FEsim

RMSE.merMod

thetaExtract

REquantile

merMod exploration functions

plotREsim

plotFEsim

draw

wiggle

subBoot

predictInterval

expectedRank

REimpact

shinyMer

See also

Author

Maintainer: Jared E. Knowles jared@civilytics.com

Authors:

Carl Frederick carlbfrederick@gmail.com

Other contributors:

Alex Whitworth whitworth.alex@gmail.com [contributor]